+++ resources/firewalld_zone.rb
@@ -35,7 +35,7 @@ property :masquerade,
          description: 'see masquerade tag in firewalld.zone(5).'
 property :ports,
          [Array, String],
-         description: 'array of port and protocol pairs. See port tag in firewalld.zone(5).',
+         description: 'array of port and protocol pairs, in `["PORT/PROTOCOL"]` format. See port tag in firewalld.zone(5).',
          coerce: proc { |o| Array(o) }
 property :priority,
          [Integer],
@@ -58,7 +58,7 @@ property :short,
          description: 'see short tag in firewalld.zone(5).'
 property :source_ports,
          [Array, String],
-         description: 'array of port and protocol pairs. See source-port tag in firewalld.zone(5).',
+         description: 'array of port and protocol pairs, in `["PORT/PROTOCOL"]` format. See source-port tag in firewalld.zone(5).',
          coerce: proc { |o| Array(o) }
 property :sources,
          [Array, String],
@@ -81,6 +81,8 @@ load_current_value do |new_resource|
     object = firewalld_service[zone_path]
     config_zone = object['org.fedoraproject.FirewallD1.config.zone']
     config_zone.getSettings2.each do |k, v|
+      # Load the current value of ports in the same format as the resource property to make it idempotent
+      v = v.map { |port, protocol| "#{port}/#{protocol}" } if %w(ports source_ports).include?(k)
       send(k, v)
     end
   else
@@ -118,6 +120,13 @@ action :update do
 
     next unless property_is_set?(property)
     new_value = new_resource.send(property)
+
+    if property == :rules_str
+      # quote the values in the rich rule just like firewalld does so it's idempotent
+      # 'rule family=ipv4 source address=192.168.0.14 accept' ➔ 'rule family="ipv4" source address="192.168.0.14" accept'
+      new_value = new_value.map { |rule| rule.gsub(/(\b\w+=)([^"\s]+)/, '\1"\2"') }
+      new_resource.rules_str = new_value
+    end
 
     if [:ports, :source_ports].include?(property)
       new_value = DBus.variant('a(ss)', new_value.map { |e| e.split('/') })