+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This file is used to list changes made in each version of the firewall cookbook.
 - Fixed: `firewall_rule` resource fails with a `--zone is an invalid option with --direct` error on firewalld.
 - Fixed: New zones created by `firewalld_zone` unexpectedly have forwarding enabled by default.
 - Fixed: `firewalld_*` resources ignore properties whose value is `false`.
+- Fixed: `firewalld_*` resources were not idempotent when using `ports`, `source_ports`, and `rich_rules` properties.
 
 ## 6.3.9 - *2024-12-05*